Crate saml_test_server

A test server for running as a SAML IdP

Example configuration file:

{
"bind_address" : "0.0.0.0",
"hostname" : "example.com",
"tls_cert_path" : "~/certs/fullchain.pem",
"tls_key_path" : "~/certs/privkey.pem",
}

Structs

AppState

ServerConfig

Functions

do_nothing
Placeholder function for development purposes, just returns a “Doing nothing” 200 response.

generate_login_form
Generate a fake login form for the user to interact with

main

saml_metadata_get
Provides a GET response for the metadata URL

saml_metadata_get_sp

saml_post_binding
Handles a POST binding

saml_redirect_get
SAML requests or responses transmitted via HTTP Redirect have a SAMLRequest or SAMLResponse query string parameter, respectively. Before it’s sent, the message is deflated (without header and checksum), base64-encoded, and URL-encoded, in that order. Upon receipt, the process is reversed to recover the original message.